lista restauracji w apce mobilnej

+++ FoodSearch.BusinessLogic.Domain.Restaurant/Models/Opinion.cs
@@ -11,6 +11,6 @@ namespace FoodSearch.BusinessLogic.Domain.Restaurant.Models
         public int Id { get; set; }
         public string UserName { get; set; }
         public string Comment { get; set; }
-        public int Rating { get; set; }
+        public short Rating { get; set; }
     }
 }
